pyproject.toml
[project]
name = "parkupine"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexander Gorokhov",email = "sashgorokhov@gmail.com"}
]
requires-python = "~3.14"
dependencies = [
    "langchain[openai] (<2)",
    "langgraph (<2)",
    "fastapi (<0.200)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "fastapi-openai-compat (==0.2)",
    "redis (>=7.2.1,<8.0.0)",
    "langchain-postgres (>=0.0.17,<0.0.18)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "sqlmodel (>=0.0.37,<0.0.38)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "langgraph-checkpoint-postgres (>=3.0.4,<4.0.0)"
]

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "langchain-tests (>=1.1.5,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "parkupine" },
]

[tool.black]
line-length = 120

[tool.pytest.ini_options]
addopts = [
    # These settings interfere with pycharms "run with coverage"...
    #    "--cov=parkupine",
    #    "--cov-report=term-missing:skip-covered",
    "--durations=5"
]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.mypy]
strict = true
disallow_untyped_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]
exclude = "^tests/"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
